tg-gambling-bot-main/requirements.txt
dotenvy==0.2.0
pyTelegramBotAPI==4.26.0
psycopg2-binary==2.9.10
